chore: rename OrganizationPanel
